修改删除接口，现在会同时删除服务器中存储的文件

+++ project_manage/views.py
@@ -146,6 +146,19 @@ def delete(request):
             print(e)
             return JsonResponse({'errno': 1003, 'msg': "不存在该项目"})
         try:
+            prototype = Prototype.objects.filter(pid=project)
+            for p in prototype:
+                delete_file(DOCUMENTS_URL + p.data)
+            document = Document.objects.filter(pid=project)
+            for d in document:
+                delete_file(DOCUMENTS_URL + d.data)
+            uml = Uml.objects.filter(pid=project)
+            for u in uml:
+                delete_file(DOCUMENTS_URL + u.data)
+        except Exception as e:
+            print(e)
+            return JsonResponse({'errno': 1005, 'msg': "删除错误"})
+        try:
             project.delete()
         except Exception as e:
             print(e)
@@ -345,6 +358,8 @@ def delete_document(request):
             doc = Document.objects.get(id=id)
         except:
             return JsonResponse({'errno': 2, 'msg': "文件不存在"})
+        if delete_file(DOCUMENTS_URL + doc.data):
+            return JsonResponse({'errno': 6668, 'msg': "文件不存在"})
         doc.delete()
         return JsonResponse({'errno': 0, 'msg': "删除成功"})
     return JsonResponse({'errno': 1, 'msg': "请求方式错误"})
